feat: setup linters
Install and configure ESLint, Prettier and Editor Config on project and set default Log Box to new Beta Version.
