+++ mobile/babel.config.js
@@ -1,3 +1,4 @@
 module.exports = {
   presets: ['module:metro-react-native-babel-preset'],
+  plugins: [['babel-plugin-root-import', { rootPathSuffix: 'src' }]],
 };
