+++ mobile/index.js
@@ -1,9 +1,7 @@
-/**
- * @format
- */
+import { AppRegistry } from 'react-native';
+import App from './src/App';
+import { name as appName } from './app.json';
 
-import {AppRegistry} from 'react-native';
-import App from './App';
-import {name as appName} from './app.json';
+require('react-native').unstable_enableLogBox();
 
 AppRegistry.registerComponent(appName, () => App);
